Use sqlite

+++ tests/conftest.py
@@ -37,13 +37,17 @@ def db(app):
 
     # insert default data
     with app.app_context():
-        _db.session.execute(text("drop database kucharka_test;"))
-        _db.session.execute(text("create schema kucharka_test;"))
-        _db.session.execute(text("use kucharka_test;"))
-        _db.session.execute(text("SET FOREIGN_KEY_CHECKS = 0;"))
-        _db.drop_all()
-        _db.create_all()
-        _db.session.execute(text("SET FOREIGN_KEY_CHECKS = 1;"))
+        if app.config["SQLALCHEMY_DATABASE_URI"] == "sqlite://":
+            _db.create_all()
+        else:
+            _db.session.execute(text("drop database kucharka_test;"))
+            _db.session.execute(text("create schema kucharka_test;"))
+            _db.session.execute(text("use kucharka_test;"))
+            _db.session.execute(text("SET FOREIGN_KEY_CHECKS = 0;"))
+            _db.drop_all()
+            _db.create_all()
+            _db.session.execute(text("SET FOREIGN_KEY_CHECKS = 1;"))
+
     return _db
